add special :manpage: mode
This mode is changing the way the directive is impacting the resulting
document. The layout changes to traditional manpage SYNOPSIS, DESCRIPTION,
etc sections. Most of the content is expected to be provided either
inline in the nested definitions (indented text below .. argparse::)
and immediately after that (for sections other than DESCRIPTION).

A number of existing sphinxarg features are not supported yet.
Specifically none of the @before @after and @replace pseudo-decorators
are handled.

Some of the new private methods are re-implementations of similar methods
from the generic code path but they could not be reused at this time.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>